add notif

+++ app/main/page.tsx
@@ -38,14 +38,27 @@ const Home: React.FC = () => {
       return;
     }
     
-    // Kirim permintaan untuk menghapus komentar dari database
-    const { error } = await supabase
+    const {data} = await supabase
+      .from('comment')
+      .select('message')
+      .eq('id', commentId)
+
+    if (!data || data.length === 0 || !data[0].message) {
+      console.error('Data is empty or message is missing');
+      return;
+    }
+
+    const { error: deleteComment } = await supabase
       .from('comment')
       .delete()
       .eq('id', commentId);
+    
+    const { error: deleteNotify } = await supabase
+      .from('notif')
+      .delete()
+      .eq('message_comment', data[0].message);
   
-    if (error) {
-      console.error('Error deleting comment:', error.message);
+    if (deleteComment || deleteNotify) {
       return;
     }
   
@@ -168,13 +181,19 @@ const Home: React.FC = () => {
 
     if (existingLikes.length > 0) {
       // If the user has already liked the post, unlike it
-      const { error } = await supabase
+      const { error: like } = await supabase
         .from('like')
         .delete()
         .eq('id_user', userId)
         .eq('id_posting', postId);
+      
+      const { error: notif } = await supabase
+        .from('notif')
+        .delete()
+        .eq('id_user', userId)
+        .eq('id_posting', postId);
 
-      if (error) {
+      if (like || notif) {
         // console.error('Error unliking post:', error.message);
         return;
       }
@@ -193,11 +212,15 @@ const Home: React.FC = () => {
       });
     } else {
       // If the user has not liked the post, like it
-      const { error } = await supabase
+      const { error: likeError } = await supabase
         .from('like')
         .insert({ id_user: userId, id_posting: postId });
+      
+      const { error: notifError } = await supabase
+        .from('notif')
+        .insert({ id_user: userId, id_posting: postId, comment: false,  like: true });
 
-      if (error) {
+      if (likeError || notifError) {
         // console.error('Error liking post:', error.message);
         return;
       }
@@ -453,15 +476,22 @@ const Home: React.FC = () => {
     const minutes = Math.floor(seconds / 60);
     const hours = Math.floor(minutes / 60);
     const days = Math.floor(hours / 24);
-  
-    if (days > 0) {
-      return `${days} day${days > 1 ? 's' : ''} ago`;
+    const months = Math.floor(days / 30);
+    const years = Math.floor(months / 12);
+    
+    if (years > 0) {
+        return `${years} year${years > 1 ? 's' : ''} ago`;
+    } else if (months > 0) {
+        return `${months} month${months > 1 ? 's' : ''} ago`;
+    } else if (days > 0) {
+        return `${days} day${days > 1 ? 's' : ''} ago`;
     } else if (hours > 0) {
-      return `${hours} hour${hours > 1 ? 's' : ''} ago`;
+        return `${hours} hour${hours > 1 ? 's' : ''} ago`;
     } else {
-      return `${minutes} minute${minutes > 1 ? 's' : ''} ago`;
+        return `${minutes} minute${minutes > 1 ? 's' : ''} ago`;
     }
   };
+
 
   const handleCommentClick = (postId: string) => {
     // Set the comment clicked ID
@@ -478,11 +508,15 @@ const Home: React.FC = () => {
     }
   
     // Mengirim komentar ke database
-    const { error } = await supabase
+    const { error: commentError } = await supabase
       .from('comment')
       .insert({ id_user: userId, id_posting: postId, message: text });
+
+    const { error: notifError } = await supabase
+      .from('notif')
+      .insert({ id_user: userId, id_posting: postId, comment: true,  like: false, message_comment: text });
   
-    if (error) {
+    if (commentError || notifError) {
       // console.error('Error sending comment:', error.message);
       return;
     }
